backend/requirements.txt
ib_insync
fastapi
uvicorn[standard]
websockets
yahooquery
pandas
numpy
scipy
requests
python-dotenv
